refactor: loaders for fields, actions and filters
