--- lib/avo/actions_loader.rb
@@ -1,11 +1,0 @@
-class Avo::ActionsLoader
-  attr_accessor :bag
-
-  def initialize
-    @bag = []
-  end
-
-  def use(klass)
-    @bag.push klass
-  end
-end
