+++ lib/avo/loaders/actions_loader.rb
@@ -1,0 +1,6 @@
+module Avo
+  module Loaders
+    class ActionsLoader < Loader
+    end
+  end
+end
