+++ lib/avo/loaders/filters_loader.rb
@@ -1,0 +1,6 @@
+module Avo
+  module Loaders
+    class FiltersLoader < Loader
+    end
+  end
+end
